# ext/bcmath: Use divide and conquer method in `bcmul`