… ext/bcmath/tests/bcmul_rec_mul.phpt