@SakiTakamachi 2024-06-02
I discovered that calculations can break under certain conditions.
This is currently only detectable in CI when testing 32bit builds.

I'm currently thinking of some test cases, and I'd like to be able to detect this even with 64-bit CI.

(I was calculating as a uint, thinking that there was no place where it would be affected, but underflow in some cases when it becomes a negative value is causing problems.)