@nielsdos 2024-06-02
You may try differential fuzzing to catch some mistakes and find potential bugs, with one fuzz variant passing inputs to the old code and the other fuzz variant passing inputs to the new code.