@Girgias 2024-06-03
The brilliant.org wiki might also be good to link for an explanation of the algorithm: https://brilliant.org/wiki/karatsuba-algorithm/