@SakiTakamachi 2024-06-03
@Girgias 
As you said, I think splitting up the PR because there are extensive bug fixes that I noticed after opening the PR.
Leave this PR as it is for now and change it in order with the new PR.