@SakiTakamachi 2024-06-03
It's too complicated, so I'll try to make the commit easier to understand. Thanks both of you for checking it out.