@SakiTakamachi 2024-06-13
I've reworked the PR so I'm closing this.

https://github.com/php/php-src/issues/14538